@@ -61,7 +61,14 @@ public abstract class ContextReferenceSetResolver<F extends PsiElement, K extend
      */
     public final Optional<K> resolve(int index) {
         Optional<K> startElement = getStartElement();
-        return startElement.isPresent() ? (texts.size() > 1 ? parseNext(startElement, texts, index) : startElement) : Optional.empty();
+
+        if (startElement.isPresent()) {
+            if (texts.size() > 1) {
+                return parseNext(startElement, texts, index);
+            }
+            return startElement;
+        }
+        return Optional.empty();
     }
 
     private Optional<K> parseNext(Optional<K> current, List<String> texts, int index) {
