@@ -1,7 +1,7 @@
 package com.baomidou.mybatisx.plugin.setting.configurable;
 
 import com.baomidou.mybatisx.plugin.setting.JavaBean2DDLSetting;
-import com.baomidou.mybatisx.plugin.ui.MyBatisGeneratorSettingPanel.MyBatisGeneratorSettingPanel;
+import com.baomidou.mybatisx.plugin.ui.MyBatisGeneratorSettingPanel;
 import com.baomidou.mybatisx.util.StringUtils;
 import com.intellij.openapi.options.ConfigurationException;
 import org.jetbrains.annotations.Nullable;
