@@ -43,7 +43,7 @@ public class BaiduTranslationService implements Translation {
     @Override
     public List<TranslationVO> toChinese(String content) {
         MultiValueMap<String, String> request = buildParameter(content, CHINESE,
-                this.appId, this.secret);
+            this.appId, this.secret);
         String response = new RestTemplateUtil().post(UNIVERSAL_API, new HashMap<>(), request, MediaType.APPLICATION_FORM_URLENCODED);
         ObjectMapper objectMapper = new ObjectMapper();
         try {
@@ -70,8 +70,8 @@ public class BaiduTranslationService implements Translation {
      *
      * @param query  需要翻译的文本
      * @param to     需要翻译的语种
-     * @param appId  appid http://api.fanyi.baidu.com/api/trans/product/desktop?req=developer
-     * @param secret secret http://api.fanyi.baidu.com/api/trans/product/desktop?req=developer
+     * @param appId  appid <a href="http://api.fanyi.baidu.com/api/trans/product/desktop?req=developer">...</a>
+     * @param secret secret <a href="http://api.fanyi.baidu.com/api/trans/product/desktop?req=developer">...</a>
      * @return
      */
     private MultiValueMap<String, String> buildParameter(String query, String to, String appId, String secret) {
@@ -89,5 +89,4 @@ public class BaiduTranslationService implements Translation {
     private String buildSign(String random, String appId, String secret, String query) {
         return MD5.md5(appId + query + random + secret);
     }
-
 }
